requirements.txt
aniso8601==8.0.0
APScheduler==3.7.0
astroid==2.4.1
attrs==19.3.0
Authlib==0.14.3
Automat==20.2.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.0
bleach==3.1.0
blinker==1.4
blis==0.4.1
cairocffi==1.1.0
CairoSVG==2.4.2
catalogue==1.0.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cobble==0.1.3
colorama==0.4.1
comtypes==1.1.7
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cssselect2==0.3.0
cymem==2.0.3
Cython==0.29.20
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-apscheduler==0.5.2
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-filter==2.2.0
django-heroku==0.3.1
django-truncate==0.1
djangorestframework==3.10.2
doc2text==0.2.4
docx==0.2.4
docx2txt==0.8
elasticsearch==7.10.1
en-core-web-sm==2.2.5
entrypoints==0.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
gunicorn==20.0.4
html2text==2020.1.16
html5lib==1.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.1
incremental==17.5.0
insultgenerator==1.0.3
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.15.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.4.3
loremipsum==1.0.5
lxml==4.5.1
mammoth==1.4.10
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mime==0.1.0
mistune==0.8.4
murmurhash==1.0.2
mysqlclient==2.0.1
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.5
notebook==5.7.8
numpy==1.18.5
oauth2==1.9.0.post1
oauthlib==2.1.0
pandas==1.0.4
pandocfilters==1.4.2
parsel==1.6.0
parso==0.4.0
pbr==5.4.0
pdfkit==0.6.1
pdfminer==20191125
pdfminer.six==20200517
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.0
prompt-toolkit==2.0.9
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoIt==0.6.0
pycparser==2.20
pycryptodome==3.9.7
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.2
pyOpenSSL==19.1.0
PyPDF2==1.26.0
Pyphen==0.9.5
pyresparser==1.0.6
pyrsistent==0.15.2
pytesseract==0.3.4
python-dateutil==2.8.0
python-docx==0.8.10
python3-openid==3.2.0
pyttsx3==2.88
pytube==9.6.0
pytube3==9.6.4
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.5.1
queuelib==1.5.0
razorpay==1.2.0
regex==2020.6.8
requests==2.22.0
requests-oauthlib==1.3.0
schedule==0.6.0
scipy==1.4.1
Scrapy==2.1.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sortedcontainers==2.2.2
soupsieve==2.0
spacy==2.2.4
SpeechRecognition==3.8.1
SQLAlchemy==1.3.20
sqlparse==0.3.0
srsly==1.0.2
stevedore==1.30.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.4.0
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.2
tqdm==4.46.1
traitlets==4.3.2
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
w3lib==1.22.0
wasabi==0.6.0
wcwidth==0.1.7
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
White-Noise==0.1.0
whitenoise==5.0.1
widgetsnbextension==3.4.2
wikipedia==1.4.0
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
